feat: Introduce dedicated structable crate
It is not possible to implement StructTable trait for Vec since trait
outside of the place where StructTable trait is defined due to the
orphan rules constraint. Reasonable possibility is to rework StructTable
trait to return headers, data row with a function combining those.
